Convert static methods to non-static
