+++ json-schema-core/src/test/resources/expected/2_constrained_int_schema.bal
